fix:修正学生选课/退课的部分逻辑。（检查stuCourse表时，某些情况应当只查status为S的数据）

+++ src/main/java/com/example/selab4/service/student/CourseService.java
@@ -19,7 +19,7 @@ public class CourseService {
     }
 
     public Response<List<CourseVO>> getAllByMajor(String studentMajor) {
-        //判定是否在选课期内
+        // 判定是否在选课期内
         String select_course_function = manager.findAdmin().getSelectcoursefunction();
         if(select_course_function.equals("off"))
             return new Response<>(Response.FAIL, "当前不是选课期", null);
@@ -42,7 +42,7 @@ public class CourseService {
             return new Response<>(Response.FAIL, "学号: " + stunum + "不存在", null);
         }
 
-        return new Response<>(Response.SUCCESS, "根据学号查找课程成功", manager.findCoursesByStudent(student));
+        return new Response<>(Response.SUCCESS, "根据学号查找课程成功", manager.findCourseVOsByStudent(student));
     }
 
     public Response<String> choose(Integer studentid, Integer courseid) {
@@ -57,41 +57,53 @@ public class CourseService {
             return new Response<>(Response.FAIL, "课程id: " + courseid + "不存在", null);
         }
 
-        // 1、课程人数上限,考虑第一次选课时不设置选课上限
-        if (!manager.checkCapacity(course,administrator)) {
+        // 1、处于选课期
+        if (!manager.chooseCourseFunctionIsOn(administrator)) {
+            return new Response<>(Response.FAIL, "不在选课期", null);
+        }
+
+        // 2、课程人数上限,考虑第一轮选课时不设置选课上限
+        if (!manager.checkCapacity(course, administrator)) {
             return new Response<>(Response.FAIL, "课程id: " + courseid + "的课程人数已满", null);
         }
 
-        // 2、学生的时间日程是否冲突
+        // 3、学生的时间日程是否冲突
         if (!manager.checkSchedule(student, course)) {
             return new Response<>(Response.FAIL, "学生id: " + studentid + "的学生上课时间发生冲突", null);
         }
 
-        // 3、学生已选/已修相同课程模板
+        // 4、学生已选/已修相同课程模板
         if(manager.courseTemplateConflict(student,course)){
             return new Response<>(Response.FAIL, "学生id: " + studentid + "的学生已经修过该课程", null);
         }
-
 
         manager.choose(student, course);
         return new Response<>(Response.SUCCESS, "学生选课成功", null);
     }
 
-    public Response<String> delete(String stunum, String coursenum) {
-        Student student = manager.findStudentByStunum(stunum);
+    public Response<String> delete(Integer studentid, Integer courseid) {
+        Student student = manager.findStudentByStudentid(studentid);
         if (student == null) {
-            return new Response<>(Response.FAIL, "学号: " + stunum + "不存在", null);
+            return new Response<>(Response.FAIL, "学生id: " + studentid + "不存在", null);
         }
-        Course course = manager.findCourseByCoursenum(coursenum);
+
+        Course course = manager.findCourseByCourseid(courseid);
         if (course == null) {
-            return new Response<>(Response.FAIL, "课程编号: " + coursenum + "不存在", null);
+            return new Response<>(Response.FAIL, "课程id: " + courseid + "不存在", null);
         }
-        // 学生是否选了该课
+
+        // 1、处于选课期
+        Administrator administrator = manager.findAdmin();
+        if (!manager.chooseCourseFunctionIsOn(administrator)) {
+            return new Response<>(Response.FAIL, "不在选课期", null);
+        }
+
+        // 2、学生是否已选该课程
         if (!manager.checkStuCourse(student, course)) {
-            return new Response<>(Response.FAIL, "学号: " + stunum + "的学生没有选过课程编号: " + coursenum + "的课程", null);
+            return new Response<>(Response.FAIL, "学号: " + student.getStunum() + "的学生没有选过课程id: " + courseid + "的课程", null);
         }
+
         manager.delete(student, course);
-        course.reduceStudent();//人数-1
         return new Response<>(Response.SUCCESS, "学生退课成功", null);
     }
